Audit PA32 follow-up: frontend memory and I/O headroom for the stress tests
pa3/tests/300-triple.t (12MB, 492k lines) timed out under a loaded
machine: the tool was ~1.7x the reference's wall time with ~13x its
system time. Three shared-frontend inefficiencies, fixed at their
owners rather than by touching the timeout:

- TranslatedChar carried two size_t provenance offsets (24 bytes per
  source code point) and the chars vector grew by doubling with no
  reserve: 432MB peak RSS and 179k minor faults on a 12MB input.
  Offsets are now uint32_t (with an explicit length_error for >= 4GB
  inputs) and TranslateSource reserves the exact upper bound.
- The four early-stage tool mains copied stdin through the
  C-stdio-synced streambuf one character at a time; they now share
  tool_stdin.h's block-fread reader (also dropping the extra
  ostringstream::str() copy of the whole input).
- CtrlExprCalculator::FinishLineResult swapped its token buffer into
  a local every line, discarding capacity; it now evaluates in place
  and clear()s. posttoken.cpp also loses its per-token endl flush,
  matching macro.cpp's documented fix.

300-triple.t: 3.6s wall / 1.1s sys / 432MB -> 1.83s / 0.13s / 168MB
(reference: 1.32s / 9.8MB), output byte-identical. Full suite
2843/2843, file audit clean; audit.md and plan.md updated.

+++ dev/src/source_translation.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -21,11 +22,16 @@ const int kInvalidChar = -2;
 // lets the tokenizer revert translations inside raw string literals by
 // rescanning the byte stream, then resuming the translated stream after
 // the closing quote.
+//
+// The offsets are 32-bit on purpose: chars is the dominant allocation of
+// every tool run (one entry per code point of the whole file), and
+// size_t offsets made the 12MB stress inputs peak at ~0.4GB. TranslateSource
+// rejects inputs that could overflow them.
 struct TranslatedChar
 {
 	int cp;
-	size_t src_begin;
-	size_t src_end;
+	uint32_t src_begin;
+	uint32_t src_end;
 };
 
 struct TranslatedSource
@@ -43,6 +49,8 @@ struct TranslatedSource
 	std::vector<TranslatedChar> chars;
 };
 
-// Applies translation phases 1 and 2 to a UTF-8 source file image. Never
-// throws: malformed bytes become kInvalidChar entries (see above).
+// Applies translation phases 1 and 2 to a UTF-8 source file image.
+// Malformed bytes never throw; they become kInvalidChar entries (see
+// above). The only throw is a length_error for inputs too large for the
+// 32-bit provenance offsets (>= 4GB).
 TranslatedSource TranslateSource(const std::string& input);
